v1.1
Multiple changes, including major things like changing the "as a sorcery" wording. That change was made mainly as it's not really reducing text length, and when it does, it ends up serving to deviate in ways that aren't always or fully relevant. Keeping the rules text and timing down at the bottom outside of specific things.